# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk CRUD Admin WA Rotator",
        "version": "1.0"
    },
    "paths": {
...
        "/api/property-admin": {
            "get": {
                "tags": [
                    "Property"
                ],
                "summary": "Retrieved List property",
                "description": "Mengambil daftar properti dengan fitur pencarian berdasarkan title, description, price, address, parking, atau kategori",
                "operationId": "aef3bf83de2d72eb9a1847e1fc570c89",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter properti berdasarkan kata kunci",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Malang"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Property retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Rumah Mewah"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Rumah mewah dengan 3 kamar tidur"
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": 1500000000
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Jl. Merdeka No.1"
                                                    },
                                                    "parking": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/property-create": {
            "post": {
                "tags": [
                    "Property"
                ],
                "summary": "Add new property",
                "description": "Menambahkan properti baru dengan detail lengkap",
                "operationId": "fe37a20aa97466f7b4c0d63d2e5b72de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "bedrooms",
                                    "bathrooms",
                                    "area",
                                    "floor",
                                    "address",
                                    "status",
                                    "category_type_id",
                                    "category_location_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Rumah Minimalis"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Rumah minimalis dengan desain modern"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 1200000000
                                    },
                                    "bedrooms": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "bathrooms": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "area": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "floor": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Kenangan No. 10"
                                    },
                                    "parking": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Available"
                                    },
                                    "category_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_location_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Property created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Property created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Rumah Minimalis"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/property-delete/{id}": {
            "delete": {
                "tags": [
                    "Property"
                ],
                "summary": "Menghapus properti",
                "description": "Delete Property",
                "operationId": "da88ee5c487b97f9c8f3999eb6189e27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID properti yang ingin dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Property deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Property not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}